+++ pyasn1/error.py
@@ -13,34 +13,6 @@ class PyAsn1Error(Exception):
     """
 
 
-class PyAsn1StringError(PyAsn1Error, UnicodeError):
-    """Create pyasn1 exception object
-
-    The `PyAsn1StringError` exception is a base class for errors relating to
-    string encoding/decoding and other related problems
-    """
-    def __init__(self, message, unicode_error=None):
-        if isinstance(unicode_error, UnicodeError):
-            UnicodeError.__init__(self, *unicode_error.args)
-        PyAsn1Error.__init__(self, message)
-
-
-class PyAsn1StringDecodeError(PyAsn1StringError, UnicodeDecodeError):
-    """Create pyasn1 exception object
-
-    The `PyAsn1StringDecodeError` exception represents a failure to decode
-    underlying bytes values to a string
-    """
-
-
-class PyAsn1StringEncodeError(PyAsn1StringError, UnicodeEncodeError):
-    """Create pyasn1 exception object
-
-    The `PyAsn1StringEncodeError` exception represents a failure to encode
-    underlying string value as bytes
-    """
-
-
 class ValueConstraintError(PyAsn1Error):
     """Create pyasn1 exception object
 
@@ -55,3 +27,33 @@ class SubstrateUnderrunError(PyAsn1Error):
     The `SubstrateUnderrunError` exception indicates insufficient serialised
     data on input of a de-serialization routine.
     """
+
+
+class PyAsn1UnicodeError(PyAsn1Error, UnicodeError):
+    """Create pyasn1 exception object
+
+    The `PyAsn1UnicodeError` exception is a base class for errors relating to
+    unicode text de/serialization.
+    """
+    def __init__(self, message, unicode_error=None):
+        if isinstance(unicode_error, UnicodeError):
+            UnicodeError.__init__(self, *unicode_error.args)
+        PyAsn1Error.__init__(self, message)
+
+
+class PyAsn1UnicodeDecodeError(PyAsn1UnicodeError, UnicodeDecodeError):
+    """Create pyasn1 exception object
+
+    The `PyAsn1UnicodeDecodeError` exception represents a failure to
+    deserialize unicode text.
+    """
+
+
+class PyAsn1UnicodeEncodeError(PyAsn1UnicodeError, UnicodeEncodeError):
+    """Create pyasn1 exception object
+
+    The `PyAsn1UnicodeEncodeError` exception represents a failure to
+    serialize unicode text.
+    """
+
+