+++ CHANGES.rst
@@ -26,6 +26,9 @@ Revision 0.4.6, released XX-06-2019
   becomes a value object.
 - Added `.reset()` method to all constructed types to turn value object
   into a schema object.
+- Added `PyAsn1UnicodeDecodeError`/`PyAsn1UnicodeDecodeError` exceptions
+  to help the caller treating unicode errors happening internally
+  to pyasn1 at the upper layers.
 - Fixed `AnyDecoder` to accept possible `TagMap` as `asn1Spec`
   to make dumping raw value operational